#ROBOMERGE-OWNER: ryan.schmidt
#ROBOMERGE-AUTHOR: ryan.schmidt
#ROBOMERGE-SOURCE: CL 16583724 via CL 16583791
#ROBOMERGE-BOT: STARSHIP (Main -> Release-Engine-Test) (v828-16531559)

[CL 16583792 by ryan schmidt in ue5-release-engine-test branch]

+++ Engine/Plugins/Experimental/MeshModelingToolset/Source/ModelingComponents/Public/Scene/MeshSceneAdapter.h
@@ -219,11 +219,11 @@ public:
 	 */
 	struct FStatistics
 	{
-		int32 UniqueMeshCount = 0;
-		int32 UniqueMeshTriangleCount = 0;
+		int64 UniqueMeshCount = 0;
+		int64 UniqueMeshTriangleCount = 0;
 
-		int32 InstanceMeshCount = 0;
-		int32 InstanceMeshTriangleCount = 0;
+		int64 InstanceMeshCount = 0;
+		int64 InstanceMeshTriangleCount = 0;
 	};
 
 	/** @return bounding box for the Actor set */
